Environment utils fix
