+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "sp-rest-proxy",
   "description": "SharePoint REST API Proxy for Node.js and Express local serve",
-  "version": "2.7.6",
+  "version": "2.7.7",
   "main": "./dist/index.js",
   "typings": "./dist/index/RestProxy",
   "scripts": {
@@ -31,30 +31,30 @@
   "homepage": "https://github.com/koltyakov/sp-rest-proxy#readme",
   "dependencies": {
     "body-parser": "^1.18.3",
-    "cors": "^2.8.4",
+    "cors": "^2.8.5",
     "express": "^4.16.4",
-    "node-sp-auth": "^2.5.6",
+    "node-sp-auth": "^2.5.7",
     "node-sp-auth-config": "^2.5.6",
     "sp-request": "^2.1.3",
     "socket.io": "^2.1.1"
   },
   "devDependencies": {
-    "@microsoft/gulp-core-build-serve": "^3.2.64",
-    "@pnp/pnpjs": "^1.2.3",
-    "@types/chai": "^4.1.6",
+    "@microsoft/gulp-core-build-serve": "^3.2.80",
+    "@pnp/pnpjs": "^1.2.6",
+    "@types/chai": "^4.1.7",
     "@types/colors": "^1.2.1",
     "@types/core-js": "^2.5.0",
     "@types/cors": "^2.8.4",
     "@types/express": "^4.16.0",
     "@types/minimist": "^1.2.0",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.0",
+    "@types/node": "^10.12.10",
     "@types/request-promise": "^4.1.42",
-    "@types/socket.io": "^1.4.38",
+    "@types/socket.io": "^2.1.0",
     "@types/socket.io-client": "^1.4.32",
     "@types/xml2js": "^0.4.3",
     "axios": "^0.18.0",
-    "chai": "^4.1.2",
+    "chai": "^4.2.0",
     "colors": "^1.3.2",
     "cpass": "^2.0.5",
     "eslint-config-standard": "^12.0.0",
@@ -65,7 +65,7 @@
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
     "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.1.3",
+    "typescript": "^3.1.6",
     "xml2js": "^0.4.19"
   }
 }
